+++ src/genie/read/spring/reorder-compress-quality-id.cc
@@ -188,7 +188,8 @@ void reorder_compress_id_pe(std::string *id_array, const std::string &temp_dir, 
         }
         genie::core::record::Chunk chunk;
         for (size_t i = 0; i < block_end[block_num] - block_start[block_num]; ++i) {
-            chunk.getData().emplace_back(1, core::record::ClassType::CLASS_U, std::move(id_array_block[i]), "", 0);
+            chunk.getData().emplace_back((uint8_t)1, core::record::ClassType::CLASS_U, std::move(id_array_block[i]), "",
+                                         (uint8_t)0);
             chunk.getData().back().addSegment(core::record::Segment("N"));
         }
         auto raw_desc = name_coder->process(chunk);
@@ -253,7 +254,7 @@ void reorder_compress_quality_pe(std::string file_quality[2], const std::string 
         for (int64_t block_num = start_block_num; block_num < end_block_num; block_num++) {
             genie::core::record::Chunk chunk;
             for (size_t i = block_start[block_num]; i < block_end[block_num]; i++) {
-                chunk.getData().emplace_back(2, core::record::ClassType::CLASS_U, "", "", 0);
+                chunk.getData().emplace_back((uint8_t)2, core::record::ClassType::CLASS_U, "", "", (uint8_t)0);
                 core::record::Segment s(std::string(quality_array[i - block_start[start_block_num]].size(), 'N'));
                 s.addQualities(std::move(quality_array[i - block_start[start_block_num]]));
                 chunk.getData().back().addSegment(std::move(s));
@@ -346,8 +347,8 @@ void reorder_compress(const std::string &file_name, const std::string &temp_dir,
             if (mode == "id") {
                 genie::core::record::Chunk chunk;
                 for (size_t i = 0; i < num_reads_block; i++) {
-                    chunk.getData().emplace_back(1, core::record::ClassType::CLASS_U,
-                                                 std::move(str_array[start_read_num + i]), "", 0);
+                    chunk.getData().emplace_back((uint8_t)1, core::record::ClassType::CLASS_U,
+                                                 std::move(str_array[start_read_num + i]), "", (uint8_t)0);
                     core::record::Segment s("N");
                     chunk.getData().back().addSegment(std::move(s));
                 }
@@ -365,7 +366,7 @@ void reorder_compress(const std::string &file_name, const std::string &temp_dir,
             } else /* mode == "quality" */ {
                 genie::core::record::Chunk chunk;
                 for (auto i = (uint32_t)start_read_num; i < start_read_num + num_reads_block; i++) {
-                    chunk.getData().emplace_back(1, core::record::ClassType::CLASS_U, "", "", 0);
+                    chunk.getData().emplace_back((uint8_t)1, core::record::ClassType::CLASS_U, "", "", (uint8_t)0);
                     core::record::Segment s(std::string(str_array[i].size(), 'N'));
                     s.addQualities(std::move(str_array[i]));
                     chunk.getData().back().addSegment(std::move(s));
